requirements.txt
beautifulsoup4==4.13.5
chromadb==1.1.0
Flask==3.1.2
konlpy==0.6.0
kss.core==1.6.5
langchain==0.3.27
langchain_chroma==0.2.6
langchain_community==0.3.29
langchain_openai==0.3.33
numpy==2.3.3
openai==1.108.2
python-dotenv==1.1.1
scikit_learn==1.7.2
selenium==4.35.0
